rag_powered_commercial_search/requirements.txt
fastapi
uvicorn
jinja2
pydantic
openai
pinecone-client
numpy
python-dotenv